removing DB_NAME env variable
