+++ common/protos/driver/driver.proto
@@ -5,9 +5,16 @@ package driver.driver;
 import "common/ack.proto";
 import "common/query.proto";
 import "common/events.proto";
+import "common/state.proto";
 
 option java_package = "com.weaver.protos.driver.driver";
 option go_package = "github.com/hyperledger-labs/weaver-dlt-interoperability/common/protos-go/driver";
+
+// Data for a View processing by dest-driver
+message WriteExternalStateMessage {
+  common.state.ViewPayload view_payload = 1;
+  common.events.ContractTransaction ctx = 2;
+}
 
 service DriverCommunication {
   // Data Sharing 
@@ -23,4 +30,8 @@ service DriverCommunication {
   // Relay uses this to get Query.requestor_signature and 
   // Query.certificate required for event subscription
   rpc RequestSignedEventSubscriptionQuery(common.events.EventSubscription) returns (common.query.Query) {}
+
+  // Events Publication
+  // the dest-relay calls the dest-driver on this end point to write the remote network state to the local ledger
+  rpc WriteExternalState(WriteExternalStateMessage) returns (common.ack.Ack) {}
 }
