+++ core/drivers/fabric-driver/.env.docker.template
@@ -9,7 +9,6 @@ DRIVER_TLS=false
 DRIVER_TLS_CERT_PATH=<path_to_tls_cert_pem_for_driver>
 DRIVER_TLS_KEY_PATH=<path_to_tls_key_pem_for_driver>
 INTEROP_CHAINCODE=<interop-chaincode-name>
-DB_NAME=driverdb
 DB_PATH=driverdbs
 DOCKER_IMAGE_NAME=ghcr.io/hyperledger-labs/weaver-fabric-driver
 DOCKER_TAG=1.4.0
