+++ core/drivers/fabric-driver/.env.template
@@ -10,5 +10,4 @@ NETWORK_NAME=network1
 DRIVER_CONFIG=
 INTEROP_CHAINCODE=interop
 MOCK=false
-DB_NAME=driverdb
 DB_PATH=driverdbs
