+++ core/relay/.env.template
@@ -2,7 +2,7 @@ PATH_TO_CONFIG=<PATH-TO-CONFIGURATION-FILE>
 RELAY_NAME=<"name" in config toml file>
 RELAY_PORT=<relay-server-port/"port" in config toml file>
 DOCKER_IMAGE_NAME=ghcr.io/hyperledger-labs/weaver-relay-server
-DOCKER_TAG=1.3.1
+DOCKER_TAG=1.4.0
 EXTERNAL_NETWORK=<docker-bridge-network>
 COMPOSE_PROJECT_NAME=<docker-compose-project-name>
 COMPOSE_PROJECT_NETWORK=<docker-compose-project-network>
